% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynadj.R
\name{comb_regr}
\alias{comb_regr}
\title{Estimate OLS from Data Frame}
\usage{
comb_regr(df)
}
\arguments{
\item{df}{A data frame or a tibble. First column of the data frame must be the \eqn{y}
variable in a regression of the type \eqn{y_i = \beta_0 + \beta_1x_{i1} + \beta_2x_{i2} + ... + \beta_nx_{in} + \epsilon_i}, for example.}
}
\value{
An object of class \code{lm}.
}
\description{
This function is just a convenient way to estimate an OLS model from a data
frame where the first variable is the dependent one of the model and the
remaining are the regressors. This function is used internally
by other functions.
}
\examples{
coreinf_br \%>\%
dplyr::select(ipca, ipcams) \%>\%
comb_regr()

coreinf_br \%>\%
 dplyr::select(ipca, ipcadp) \%>\%
 dplyr::mutate(`ipca(-1)` = dplyr::lag(ipca)) \%>\%
 comb_regr()
}
