% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynadj.R
\name{dyn_adj}
\alias{dyn_adj}
\title{Map over Several Horizons to Extract Dynamic Adjustment Parameter}
\usage{
dyn_adj(inf, core, H, p, ...)
}
\arguments{
\item{inf}{A vector containing headline inflation.}

\item{core}{A vector containing a measure of core inflation.}

\item{H}{An integer that gives the maximum horizon.}

\item{p}{An integer, lag in the regression (see details).}

\item{...}{Additional parameters.}
}
\value{
A tibble with the dynamic adjustment parameter and p-value for the t test
of \eqn{\lambda^j_h = 0} for \eqn{j = ( , *)}. Row 1 of
the tibble is \eqn{h = 1}, row 2 is \eqn{h = 2} and so on. For estimating
\eqn{\lambda^*_h}, put \code{core} as the first argument of the function
and headline inflation as the second.
}
\description{
\code{dyn_adj} is a fast way to extract the parameter
of the dynamic adjustment for a range of horizons.
}
\details{
For a model \eqn{\pi_{t + h} - \pi_t =
a_0 + \lambda_h(\pi_t - \pi^*_t) + \sum_{i = 1}^pa_i\pi_{t - i} +
e_{t + h}}, where \eqn{\pi} is the headline inflation and
\eqn{\pi^*} a measure of core inflation,
this function gives \eqn{\lambda_h} for \eqn{h \in 1, 2 ..., H}.

In core inflation's literature, parameters \eqn{\lambda_h}
and \eqn{\lambda^*_h} of the following regressions are important:
\itemize{
\item \eqn{\pi_{t + h} - \pi_t = a_0 + \lambda_h(\pi_t - \pi^*_t) +
\sum_{i = 1}^pa_i\pi_{t - i} + e_{t + h}},
\item \eqn{\pi^*_{t + h} - \pi^*_t = a^*_0 + \lambda^*_h(\pi_t - \pi^*_t) +
\sum_{i = 1}^pa^*_i\pi^*_{t - i} + e^*_{t + h}}.
}
A good core inflation measure should imply \eqn{\lambda_h < 0} and
\eqn{\lambda^*_h = 0}.
}
\examples{
inf_head <- coreinf_br[["ipca"]]
inf_corems <- coreinf_br[["ipcams"]]
inf_coredp <- coreinf_br[["ipcadp"]]
dyn_adj(inf_head, inf_corems, 4, 5, recursive = TRUE)
out <- purrr::map(list(inf_corems, inf_coredp), ~ dyn_adj(inf_head, .x, 4, 5, recursive = TRUE))
purrr::map(out, dplyr::summarise_all, mean)
}
\seealso{
\code{\link{dyn_adj_est}, \link{dyn_adj_best},
\link{dyn_adj_pred}}
}
