% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynadj.R
\name{dyn_adj_best}
\alias{dyn_adj_best}
\title{Select Best Dynamic Adjust Model}
\usage{
dyn_adj_best(models, ic = BIC)
}
\arguments{
\item{models}{An object generated by the funcion \code{\link{dyn_adj_est}}.}

\item{ic}{Information criterion BIC or AIC.}
}
\value{
A list containing the following elements:
\item{data}{A tibble with the data used in fitting the model.}
\item{model}{An \code{lm} object.}
\item{n_obs}{Number of observations.}
\item{h}{Horizon used in direct estimation. See \code{\link{dyn_adj_est}}.}
}
\description{
Given models estimated by the function \code{\link{dyn_adj_est}}, \code{dyn_adj_best} selects
the best one.
}
\examples{
inf_head <- coreinf_br[["ipca"]]
inf_corems <- coreinf_br[["ipcams"]]
models <- dyn_adj_est(inf_head, inf_corems, 2, 2)
dyn_adj_best(models, ic = AIC)
}
\seealso{
\code{\link{dyn_adj_est}}
}
