% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynadj.R
\name{dyn_adj_pred}
\alias{dyn_adj_pred}
\title{Extract the Dynamic Adjustment Parameter}
\usage{
dyn_adj_pred(best_model)
}
\arguments{
\item{best_model}{An object from function \code{\link{dyn_adj_best}}.}
}
\value{
A list of two elements: dynamic adjustment parameter and
p-value.
}
\description{
This function is a convenient way of extracting both the dynamic adjustment parameter
and the associated p-value from the best dynamic model estimated by the
function \code{\link{dyn_adj_best}}.
}
\examples{
inf_head <- coreinf_br[["ipca"]]
inf_corems <- coreinf_br[["ipcams"]]
models <- dyn_adj_est(inf_head, inf_corems, 2, 2)
best_model <- dyn_adj_best(models, ic = AIC)
dyn_adj_pred(best_model)
}
\seealso{
\code{\link{dyn_adj_est}, \link{dyn_adj_best}}
}
