% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe.R
\name{ebayesthresh.wavelet}
\alias{ebayesthresh.wavelet}
\title{Empirical Bayes thresholding on the levels of a wavelet transform}
\description{
See \code{EbayesThresh::\link[EbayesThresh]{ebayesthresh.wavelet}} for details.
}
\keyword{internal}
