% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecastOOP.R
\name{error_wave_summary}
\alias{error_wave_summary}
\title{Summarize Error Prediction Wavelet Models}
\usage{
error_wave_summary(h, ...)
}
\arguments{
\item{h}{A positive integer that controls which maximum horizon is being used in forecasting (\eqn{1, 2, ..., h}).}

\item{...}{Additional parameters to pass to the function
\code{\link{model_error_wave}}.}
}
\value{
A tibble with the average forecast errors for all horizon \code{h} considered and
for all  wavelet models estimated.
}
\description{
This function calculates the average forecast errors over forecasting horizons.
Those errors are estimated by the function \code{\link{model_error_wave}}.
}
\examples{
wthr_wd <- list(
  filter.number = 2:4,
  type = c("wavelet", "station")
)

wthr_thr <- list(
  type = c("soft", "hard"),
  policy = c(
    "universal",
    "BayesThresh"
  )
)

wthr_args <- wav_args_wthr(wthr_wd, wthr_thr, 3:5)

pq <- lags(2, 1)

inf_head <- coreinf_br[["ipca"]]

error_wave_summary(2, x = wthr_args, y = inf_head, lags = pq, k = 3, RMSE = TRUE)
}
\seealso{
\code{\link{model_error_wave}}
}
