% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgsdata.R
\name{getsgs}
\alias{getsgs}
\title{Extract Data from SGS Database}
\usage{
getsgs(namex, numberx)
}
\arguments{
\item{namex}{A vector of strings containing the name of the series.}

\item{numberx}{An integer vector representing the code number of the series in SGS database.}
}
\value{
Tibble
}
\description{
Extract data from Central Bank of Brazil statistical database \href{https://www3.bcb.gov.br/sgspub/localizarseries/localizarSeries.do?method=prepararTelaLocalizarSeries}{Time Series Management System (SGS)}.
}
\examples{
\dontrun{
# 433  ipca (cpi)
# 11427 ex0 (core inflation)
n <- c(433, 11427)
x <- c("ipca", "ex0")
# Sys.setenv(no_proxy = "bcb.gov.br") # configuration for internal bcb connection
inf <- getsgs(x, n)
}
}
