% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lags.R
\name{lags}
\alias{lags}
\title{Lag Index Combination}
\usage{
lags(p, q, recursive = FALSE)
}
\arguments{
\item{p}{Integer, maximum lag \code{p}.}

\item{q}{Integer, maximum lag \code{q}.}

\item{recursive}{Logical, \code{FALSE} or \code{TRUE} (see details).}
}
\value{
Data frame
}
\description{
Generates a data frame with pairs of indexes given the maximum lags
of both \code{p} and \code{q}. This function is useful as an intermediate step for
constructing lag variables to be used in regression models.
}
\details{
Condiser the following regression:
\itemize{
\item \eqn{y_{i,t} = a_0 + y_{i,t-pstart} + ... + y_{i,t-pend} + x_{i,t-qstart} + ... + x_{i,t-qend} + e_{i,t}}.
}
\code{lags} constructs a data frame with indexes \eqn{pstar}, \eqn{pend},
\eqn{qstart} and \eqn{qend}. This is an intermediate step for obtaining
the variables to estimate a lot of possible models.
If \code{recursive = FALSE}, both \eqn{pstart} and \eqn{qstart} are always = 1
and \eqn{pend} and \eqn{qend} grows from one to \code{p} and \code{q}, respectively.
If \code{recursive = TRUE}, on the other hand, \eqn{pstart} and \eqn{qstart}
also grows from one to \code{p} and \code{q}. This last case is
appropriated if one wants to construct variables for
estimating models with more combinations of lags.
The cost of this option is that is time-consuming in
terms of computation.
}
\examples{
lags(2, 1, recursive = FALSE)
}
