% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mavc.R
\name{mavc}
\alias{mavc}
\title{Centered Moving Avarage}
\usage{
mavc(x, k)
}
\arguments{
\item{x}{Numeric vector or a time series.}

\item{k}{Integer, window length.}
}
\value{
A numeric vector.
}
\description{
\code{mavc} calculates the \code{k}-period centered moving avarage of a vector or time series.
When \code{k} is even the series length is increased by one and the two
endpoints are weighted by 0.5.
}
\examples{
coreinf_br \%>\%
dplyr::select(-date) \%>\%
purrr::map_df(~mavc(.x, 25)) \%>\%
tibble::add_column(date = coreinf_br$date, .before = 1)
}
