% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecastOOP.R
\name{model_best}
\alias{model_best}
\title{Select Best Model}
\usage{
model_best(models, ic = BIC)
}
\arguments{
\item{models}{An object generated by the funcion
\code{\link{model_est}} or \code{\link{model_est_single}}.}

\item{ic}{Information criterion: \code{BIC} or \code{AIC}.}
}
\value{
A list containing the following elements:
\item{data}{Tibble with the data used in fitting the model.}
\item{model}{An \code{lm} object.}
\item{n_obs}{Number of observations of the dependent variable.}
\item{h}{Horizon used in direct estimation.}
}
\description{
A function for selecting the best model fitted by the
function \code{\link{model_est}} or \code{\link{model_est_single}}.
}
\examples{

inf_head <- coreinf_br[["ipca"]]
inf_corems <- coreinf_br[["ipcams"]]
pq <- lags(2, 1)
model1 <- model_est(inf_head, inf_corems, 2, pq)
model_best(model1)

p <- unique(lags(2, 1)[, 1:2])
model2 <- model_est_single(inf_head, 2, p)
model_best(model2, ic = AIC)
}
\seealso{
\code{\link{model_est}}, \code{\link{model_est_single}}
}
