% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecastOOP.R
\name{model_error}
\alias{model_error}
\title{Error Prediction Two Variables}
\usage{
model_error(k, y, xreg, lags, h, RMSE = TRUE)
}
\arguments{
\item{k}{An integer to determine the number of observations to include
in the out-of-sample set.}

\item{y}{A numeric vector or time series of the dependet variable; tipplically
the headline inflation.}

\item{xreg}{A numeric vector or time series of the independent variable; a measure
of core inflation that is never revised.}

\item{lags}{A data frame generated by the function \code{\link{lags}}.}

\item{h}{An integer to set the desired horizon.}

\item{RMSE}{Logical: if \code{TRUE} returns \code{RMSE}, if \code{FALSE} returns \code{MAE}.}
}
\value{
An integer contaning the calculated error.
}
\description{
This function calculates the root mean square error or mean absolute error of
models estimated by the function \code{\link{model_fcast}}. This function
is appropriate for forecasting evaluation of a
core inflation measure that is never revised.
}
\examples{
inf_head <- coreinf_br[["ipca"]]
inf_corems <- coreinf_br[["ipcams"]]
pq <- lags(2, 1)
model_error(24, inf_head, inf_corems, pq, 2, RMSE = TRUE)
purrr::map_dbl(1:12, ~ model_error(24, inf_head, inf_corems, pq, .x, RMSE = FALSE))
}
\seealso{
\code{\link{model_est}}, \code{\link{model_best}},
\code{\link{model_fcast}}.
}
