% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecastOOP.R
\name{model_error_single}
\alias{model_error_single}
\title{Error Prediction Single Variable}
\usage{
model_error_single(k, y, lags, h, RMSE = TRUE)
}
\arguments{
\item{k}{An integer to determine the number of observations to include
in the out-of-sample set.}

\item{y}{A numeric vector or time series of the dependet variable; tipplically
the headline inflation.}

\item{lags}{A data frame generated by the function \code{\link{lags}}.}

\item{h}{An integer to set the desired horizon.}

\item{RMSE}{Logical: if \code{TRUE} returns \code{RMSE}, if \code{FALSE} returns \code{MAE}.}
}
\value{
An integer contaning the calculated error.
}
\description{
Calculates the root mean square error or mean absolute error of
univariate models estimated by the function \code{\link{model_fcast}} or
\code{\link{model_est_single}}.
}
\examples{
inf_head <- coreinf_br[["ipca"]]
pq <- lags(2, 1)
purrr::map_dbl(1:12, ~ model_error_single(24, inf_head, unique(pq[, 1:2]), .x, RMSE = FALSE))
}
\seealso{
\code{\link{model_est_single}}, \code{\link{model_best}},
\code{\link{model_fcast}}.
}
