% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecastOOP.R
\name{model_error_wave}
\alias{model_error_wave}
\title{Error Prediction Wavelet Models}
\usage{
model_error_wave(x, y, h, lags, k, RMSE = TRUE)
}
\arguments{
\item{x}{An object of class \code{args_wshr}, \code{args_ebthr} or \code{args_wthr}.}

\item{y}{A numeric vector or a time series for wavelet-based signal estimation.
Headline inflation for estimation of a wavelet core inflation measure.}

\item{h}{An integer to set the desired horizon.}

\item{lags}{A data frame generated by the function \code{\link{lags}}.}

\item{k}{An integer to determine the number of observations to include in the out-of-sample set.}

\item{RMSE}{logical: if \code{TRUE} returns \code{RMSE}, if \code{FALSE} returns \code{MAE}.}
}
\value{
A vector of errors for every wavelet model considered.
}
\description{
\code{model_error_wave} estimates the prediction error of models
fitted by the function \code{\link{model_est}} when the \code{x} variable
is a wavelet-based signal estimation of \code{y}.
For each new observation \code{k} included in the
out-of-sample set, the wavelet-based signal is reestimated.
This function supports
several specifications of wavelet models being a generalization
of \code{\link{model_error_wave_single}}.
}
\examples{
wthr_wd <- list(
  filter.number = 4,
  type = c("wavelet", "station")
)

wthr_thr <- list(
  type = c("soft", "hard"),
  policy = c(
    "universal",
    "BayesThresh"
  )
)

wthr_args <- wav_args_wthr(wthr_wd, wthr_thr, 4:5)

h <- 1:3
names(h) <- paste0("h", h)

pq <- lags(2, 1)

inf_head <- coreinf_br[["ipca"]]

# h = 2

model_error_wave(wthr_args, inf_head, h = 2, pq, 12, RMSE = TRUE)

# h = 1:3 (mean of the errors for each wavelet model estimated)
\dontrun{

pred_wthr <- purrr::map_df(
  h,
  ~ model_error_wave(wthr_args, inf_head, .x, pq, 12, RMSE = TRUE)
) \%>\%
  dplyr::mutate(pred = purrr::possibly(rowMeans, NA)(.)) \%>\%
  dplyr::select(pred)

pred_wthr
}

}
\seealso{
\code{\link{wav_smooth}}, \code{\link{model_est}},
\code{\link{model_best}}, \code{\link{model_fcast}}
}
