% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecastOOP.R
\name{model_error_wave_single}
\alias{model_error_wave_single}
\title{Error Prediction for Single Wavelet Model}
\usage{
model_error_wave_single(k, y, lags, h, RMSE = TRUE, ...)
}
\arguments{
\item{k}{An integer to determine the number of observations to include in the out-of-sample set.}

\item{y}{A numeric vector or a time series for wavelet-based signal estimation.
Headline inflation for estimation of a wavelet core inflation measure.}

\item{lags}{A data frame generated by the function \code{\link{lags}}.}

\item{h}{An integer to set the desired horizon.}

\item{RMSE}{logical: if \code{TRUE} returns \code{RMSE}, if \code{FALSE} returns \code{MAE}.}

\item{...}{Additional parameters to pass to function \code{\link{smooth_wavelet}}.}
}
\value{
A numeric vector of length one.
}
\description{
\code{model_error_wave_single} estimates the prediction
error of models fitted by the
function \code{\link{model_est}} when the \code{x} variable is
\code{y} smoothed by wavelet models. For each new observation
\code{k} inlucluded in the out-of-sample set,
the wavelet-based signal is reestimated.
This function computes
the prediction error for a single wavelet specification and
therefore is similar to \code{\link{model_error_wave}}
that is used for computing the error for several specifications of
wavelet models.
}
\examples{
inf_head <- coreinf_br[["ipca"]]
pq <- lags(2, 1)

# h = 2
model_error_wave_single(
36, inf_head, pq, 2, RMSE = FALSE,
thfun = ebayesthresh.wavelet,
wtfun = "dwt",
wtfunlist = list(wf = "d6", n.levels = 4),
a = 0.5, vscale = "level", prior = "cauchy")

# h = 1:6; k = 20

purrr::map_dbl(1:6, ~ model_error_wave_single(
20, inf_head, pq, .x, RMSE = FALSE,
thfun = ebayesthresh.wavelet,
wtfun = "dwt",
wtfunlist = list(wf = "d6", n.levels = 4),
a = 0.5, vscale = "level", prior = "cauchy"))
}
\seealso{
\code{\link{smooth_wavelet}, \link{model_error_wave}},
\code{\link{model_fcast}, \link{model_best}, \link{model_est}}
}
