% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecastOOP.R
\name{model_est}
\alias{model_est}
\title{Estimate Direct Model}
\usage{
model_est(y, xreg, h, lags)
}
\arguments{
\item{y}{A numeric vector or time series.}

\item{xreg}{A numeric vector for the exogenous variable.}

\item{h}{An integer. The horizon of prediction.}

\item{lags}{A data frame generated by the function \code{\link{lags}}.}
}
\value{
A list with following components:
\item{model}{A \code{lm} object.}
\item{n_obs}{Number of observations of \code{y}.}
\item{h}{The horizon over which y is to be forecast.}
}
\description{
Given \code{y} and \code{xreg}, this function estimates direct models as
described bellow in the details section.
}
\details{
Condiser the following regression:
\itemize{
\item \eqn{y_{i,t+h} = a_0 + y_{i,t + 1 - pstart} + ... + y_{i,t + 1 - pend} + x_{i,t + 1 - qstart} + ... + x_{i,t + 1 - qend} + e_{i,t + h}}.
}
\code{model_est} estimates the above regression for all indexes in the
data frame \code{lags} (\code{p_start} \eqn{\ge 1} and \code{q_start} \eqn{\ge 1}). Forecasting from a model with this
structure is called direct forecast, where \eqn{h}
denotes the horizon over which \eqn{y} is to be forecast. This
approach can be used to forecast several steps ahead.

A measure of core inflation (\eqn{\pi^*}) should
improve forecasting of headline inflation (\eqn{\pi}). A way of test
this property is running a regression like that one above:
\itemize{
\item \eqn{\pi_{t+h} = \alpha_0 + \sum_{i = 1}^p\beta_i\pi_{t-i+1} +
 \sum_{i = 1}^q\gamma_i\pi^*_{t-i+1} e_{t + h}}.
}
Using \code{recursive = FALSE} (default) in function \code{\link{lags}} is
equivalent to estimate the previous equation for headline inflation.
If \code{recursive = TRUE},
\eqn{i} is not necessarily initiated in 1, i.e, \code{p_start}
and \code{q_start}  also changes in
the first equation. Allowing for \code{p_start} and \code{q_start} to vary
implies more models estimated because it opens up
more possibilities.
}
\examples{
inf_head <- coreinf_br[["ipca"]]
inf_corems <- coreinf_br[["ipcams"]]
pq <- lags(2, 1)
model_est(inf_head, inf_corems, 2, pq)
}
\seealso{
\code{\link{model_est_single}} \code{\link{model_fcast}}
}
