% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecastOOP.R
\name{model_fcast}
\alias{model_fcast}
\title{Forecast from the Best Model}
\usage{
model_fcast(best_model)
}
\arguments{
\item{best_model}{A model generated by the function \code{\link{model_best}}.}
}
\value{
A prediction value for the horizon \code{h} considered in the
function \code{\link{model_best}}.
}
\description{
\code{model_fcast} is a function for forecasting \code{h} step ahead from a model selected by
the function \code{\link{model_best}}.
}
\examples{
inf_head <- coreinf_br[["ipca"]]
inf_corems <- coreinf_br[["ipcams"]]
pq <- lags(2, 1)
model1 <- model_est(inf_head, inf_corems, 2, pq)
best <- model_best(model1)
model_fcast(best)
}
\seealso{
\code{\link{model_best}}
}
