% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bias.R
\name{no_bias}
\alias{no_bias}
\title{Bias}
\usage{
no_bias(y, x)
}
\arguments{
\item{y}{A numeric vector or \code{ts} (inflation, \eqn{\pi})}

\item{x}{A numeric vector or \code{ts }(core inflation, \eqn{\pi^*})}
}
\value{
A (1 x 4) tibble.
}
\description{
Estimates a regression model of
\code{y} (healine inflation, \eqn{\pi})
in \code{x} (core inflation, \eqn{\pi^*}),
\itemize{
\item \eqn{\pi_{it} = a + b\pi^*_{it} + e_{it}},
} and
gives statistics for \eqn{a, b, \overline{R}^2} and F-test \eqn{H0:
a = 0, b = 1}. \code{no_bias} is useful to find out
whether or not the bias in a core inflation measure (\eqn{\pi^*}) is
significant.
}
\examples{
inf_head <- coreinf_br[["ipca"]]
inf_corems <- coreinf_br[["ipcams"]]
no_bias(inf_head, inf_corems)
}
\seealso{
\code{\link{no_bias2}}
}
