% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bias.R
\name{no_bias2}
\alias{no_bias2}
\title{Bias p-value}
\usage{
no_bias2(y, x)
}
\arguments{
\item{y}{A numeric vector or \code{ts} (inflation, \eqn{\pi})}

\item{x}{A numeric vector or \code{ts }(core inflation, \eqn{\pi^*})}
}
\value{
A numeric vector of length one.
}
\description{
Extracts p-value from the model estimated by the function
\code{\link{no_bias}}. If the mean of \code{y} is equal of  \code{x}, p-value is
expected to be higher than 0.05. In this circuntance, bias of
the core inflation measure is not significant. See \code{\link{no_bias}}
for more information.
}
\examples{
inf_head <- coreinf_br[["ipca"]]
inf_corems <- coreinf_br[["ipcams"]]
no_bias2(inf_head, inf_corems)
}
\seealso{
\code{\link{no_bias}}
}
