% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{prior_a}
\alias{prior_a}
\title{Determine Scale Factor Parameter "a" for Laplace Distribution}
\usage{
prior_a(
  x,
  h,
  k,
  lags,
  interval = c(0.1, 5),
  tol = 0.01,
  wtfun = "dwt",
  wtfunlist = list(),
  ...
)
}
\arguments{
\item{x}{A vector or a time series containing the data (headline inflation).}

\item{h}{An integer to set the desired horizon.}

\item{k}{An integer to determine the number of observations to include
in the out-of-sample set.}

\item{lags}{A data frame generated by the function \code{\link{lags}}.}

\item{interval}{A vector containing the end-points of the interval to be searched for the minimum. See: \code{stats::\link[stats]{optimize}}.}

\item{tol}{The desired accuracy. See: \code{stats::\link[stats]{optimize}}.}

\item{wtfun}{A wavelet transform function: \code{waveslim::\link[waveslim]{dwt}} or \code{waveslim::\link[waveslim]{modwt}}.}

\item{wtfunlist}{A named list of parameters to pass to the function \code{\link[=dwt]{dwt}} or \code{\link[=modwt]{modwt}}.}

\item{...}{Additional parameters to pass to the function \code{\link{smooth_wavelet}}.}
}
\value{
A numeric vector of length one.
}
\description{
The function \code{prior_a} determines the parameter
\code{a} of the laplace distribution that is used in
\link[=ebayesthresh.wavelet]{empirical
Bayes thresholding approach}. Estimation  is based on
minimizing the avarage of forecast error.
}
\examples{
pq <- lags(2, 1)
ipca <- coreinf_br$ipca
prior_a(
  x = ipca, h = 2, k = 15,
  lags = pq, interval = c(0.1, 3),
  tol = 0.2, wtfun = "dwt",
  wtfunlist = list(
    wf = "haar", n.levels = 4,
    boundary = "reflection"
  ),
  vscale = "independet", threshrule = "mean"
)
}
