% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmsemae.R
\name{rmse_mae}
\alias{rmse_mae}
\title{Map RMSE and MAE by Combining Two Data Frames}
\usage{
rmse_mae(dfa, dfb)
}
\arguments{
\item{dfa}{A data frame of numeric vectors.}

\item{dfb}{A data frame of numeric vectors. The number
of rows of \code{dfb} must be equal to \code{dfa}.}
}
\value{
A data frame
}
\description{
\code{rmse_mae} is a functional to compute root mean square error and mean absolute
error for all variables in a data frame with respect to all variables in
another data frame. The number of rows of both data
frames must be equal.
}
\examples{
# Measuring how far wavelet core inflation measures are
# from moving averages of headline inflation.

inf_head <- coreinf_br[["ipca"]]

mavc_order <- c(13, 25, 37)
names(mavc_order) <- paste0("mavc", mavc_order)
df_mavc <- purrr::map_df(mavc_order, ~mavc(x = inf_head, k = .))

corewav1 <- wmtsa::wavShrink(inf_head)
corewav2 <- wmtsa::wavShrink(inf_head, wavelet = "haar")
df_core <- tibble::tibble(corewav1, corewav2)

rmse_mae(df_core, df_mavc)
}
