% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wavcore.R
\name{wav_args_ebthr}
\alias{wav_args_ebthr}
\title{Wavelet EbayesThresh Arguments}
\usage{
wav_args_ebthr(xform, wtlist, ebthrlist)
}
\arguments{
\item{xform}{A named list of wavelet transform methods. Types can be \code{"dwt"} and \code{"modwt"}.}

\item{wtlist}{A named list of arguments to use in the wavelet
decomposition, excluding the input vector \code{x}, performed by the
functions \code{waveslim::\link[waveslim]{dwt}} or \code{waveslim::\link[waveslim]{modwt}}.}

\item{ebthrlist}{A named list of arguments that can be passed to the function \cr
\code{EbayesThresh::\link[EbayesThresh]{ebayesthresh.wavelet}} excluding
\code{xtr}.}
}
\value{
S3 objetc of class args_ebthr. This object is a list of
the following elements:
\describe{
\item{args_tbl}{A tibble with arguments to get wavevel-based
signal estimation from \cr \code{EbayesThresh::\link[EbayesThresh]{ebayesthresh.wavelet}}}.
\item{nwt_args}{A integer vector of length two. The first
element is the number of arguments to pass to the function
 \code{waveslim::\link[waveslim]{dwt}} or \code{waveslim::\link[waveslim]{modwt}}. The
 second element is the number of arguments to the  \code{EbayesThresh::\link[EbayesThresh]{ebayesthresh.wavelet}}}.
}
}
\description{
The main goal of this function is to generate a tibble
with combinations of some possible arguments that can be passed to
get wavelet-based signal estimation from the function \cr
\code{EbayesThresh::\link[EbayesThresh]{ebayesthresh.wavelet}}.
}
\examples{

# Wavelet transform

ebthr_xform <- list(wt = c("dwt", "modwt"))

# Wavelet filter (package waveslim)

ebthr_wt <- list(
  wf = c("haar", "d4", "d6", "la8"),
  n.levels = 1:4,
  boundary = c("periodic", "reflection")
)

# Some arguments that can be passed to EbayesThresh::ebayesthresh.wavelet

ebthr_ebwav <- list(
  vscale = c("level", "independet"),
  a = 0.5,
  prior = c("laplace", "cauchy"),
  threshrule = c("median", "mean", "soft", "hard")
)

ebthr_args <- wav_args_ebthr(ebthr_xform, ebthr_wt, ebthr_ebwav)
}
\seealso{
\code{\link{wav_smooth}}
}
