% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wavcore.R
\name{wav_args_wshr}
\alias{wav_args_wshr}
\title{Wavelet Shrinkage Arguments}
\usage{
wav_args_wshr(wtlist)
}
\arguments{
\item{wtlist}{A named list of wavelet arguments for getting
wavelet-based signal estimation from \code{wmtsa::\link[wmtsa]{wavShrink}}.}
}
\value{
S3 objetc of class args_wshr. This object is a list of
the following element:
\describe{
\item{args_tbl}{A tibble with arguments to get wavevel-based
signal estimation from \cr \code{wmtsa::\link[wmtsa]{wavShrink}}}.
}
}
\description{
A tibble with combinations of some possible arguments
that can be passed to get wavelet-based signal
estimation from the function \code{wmtsa::\link[wmtsa]{wavShrink}}.
}
\examples{
# wmtsa arguments for using with wmtsa package
wshr_wmtsa <- list(
  wavelet = c("haar", "d4", "d6", "d8", "s8"),
  n.level = 1:4,
  shrink.fun = c("hard", "soft", "mid"),
  thresh.fun = c("universal", "minimax", "adaptive"),
  xform = c("dwt", "modwt"),
  reflect = c(TRUE, FALSE)
)
wav_args_wshr(wshr_wmtsa)
}
\seealso{
\code{\link{wav_smooth}}
}
