% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wavcore.R
\name{wav_args_wthr}
\alias{wav_args_wthr}
\title{Wavelet Wavethresh Arguments}
\usage{
wav_args_wthr(wdlist, thrlist, lr)
}
\arguments{
\item{wdlist}{A named list of arguments for wavelet transform by using
\code{wavethresh::\link[wavethresh]{wd}}, excluding the
\code{data} argument.}

\item{thrlist}{A named list of arguments that can be used in function
\code{wavethresh::\link[wavethresh]{threshold}}, excluding the \code{wd} argument.}

\item{lr}{An integer vector which determines the minimum scale
level that is thresholded. The maximum scale level is given by
\code{wavethresh::\link[wavethresh]{nlevelsWT}(wd)-1}. For example,
\code{lr[1] = 3} is equivalent to \code{levels = 3:(nlevelsWT(wd) - 1)}
in \code{wavethresh::\link[wavethresh]{threshold.wd}}.}
}
\value{
S3 objetc of class args_wthr. This object is a list of
the following elements:
\describe{
\item{args_tbl}{A tibble with arguments to get wavevel-based
signal estimation from \cr \code{wavethresh::\link[wavethresh]{threshold}}.}
\item{nwt_args}{A integer vector of length two. The first
element is the number of arguments to pass to the function
 \code{wavethresh::\link[wavethresh]{wd}}, excluding \code{data}. The
 second element is the number of arguments to the
 \code{wavethresh::\link[wavethresh]{threshold}} function, excluding \code{wd}.}
}
}
\description{
The main goal of this function is to generate a tibble
with combinations of some possible arguments that can be passed to
get wavelet-based signal estimation from the function
\code{wavethresh::\link[wavethresh]{threshold}}.
}
\examples{
wthr_wd <- list(
  filter.number = 1:4,
  family = c("DaubLeAsymm", "DaubExPhase"),
  type = c("wavelet", "station"),
  bc = c("periodic", "symmetric")
)

wthr_thr <- list(
  type = c("soft", "hard"),
  policy = c(
    "universal",
    "sure",
    "BayesThresh", "cv"
  ),
  boundary = c(TRUE, FALSE)
)

wthr_args <- wav_args_wthr(wthr_wd, wthr_thr, 3:5)
}
\seealso{
\code{\link{wav_smooth}}
}
