% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{wcore_spec}
\alias{wcore_spec}
\title{Extract Wavelet Denoised Series}
\usage{
wcore_spec(x, filterexpr)
}
\arguments{
\item{x}{An object generated by the functions
\code{\link{wcore_summary}} or \code{\link{wcore_summary_fcast}}.}

\item{filterexpr}{A logical condition involving
variables  used in functions of type \code{\link{wav_args}}.}
}
\value{
A tibble
}
\description{
This function computes the median series
representing the wavelet-based signal estimation
 series for a specific condition.
}
\examples{
library(lubridate)

inf_head <- coreinf_br[["ipca"]]
date_start <- coreinf_br[["date"]][1]
ts_start <- c(year(date_start), month(date_start))
inf_head_ts <- ts(inf_head, start = ts_start, frequency = 12)

wshr_obj <- wav_args_wshr(list(
wavelet = c("haar", "d4", "d6", "d8", "s8"),
n.level = 1:4
))

core_wavelet <- wav_smooth(inf_head, wshr_obj)

core_wavelet2 <- wcore_table(core_wavelet, inf_head_ts)

x <- wcore_summary(core_wavelet2, inf_head)

wcore_spec(x, wavelet == "haar")
}
