% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{wcore_summary}
\alias{wcore_summary}
\title{Summarize Wavelet Core Inflation}
\usage{
wcore_summary(x, y)
}
\arguments{
\item{x}{An object generated by the function \code{\link{wcore_table}}.}

\item{y}{A numeric vector (headline inflation) that must be the same \code{x} used in \code{\link{wav_smooth}}.}
}
\value{
A tibble. New columns are added to the object \code{x}:
\describe{
\item{sd}{Standard desviation of \code{x_sm} column that represents the
wavelet core inflation measure.}
\item{mav25}{RMSE of headline inflation 25-period centered
 moving avarage (\code{\link{mavc}}) and
 wavelet core inflation.}
 \item{no_bias}{p-value calculated by the
 function \code{\link{no_bias2}} to
 verify if the bias of the wavelet core inflation
 measure is not significant.}
 \item{alpha, val_p, alpha1 and val_p1}{Coefficients
 \eqn{\lambda_h} and \eqn{\lambda^*_h} and respective
 p-value from function \code{\link{dyn_adj}}, where
 \code{H = 1 and p = 1}.}
}
}
\description{
This function calculates statistics that summarize the quality of
an estimated wavelet-based core inflation measure, excluding the forecasting
criterion.
}
\examples{
library(lubridate)

inf_head <- coreinf_br[["ipca"]]
date_start <- coreinf_br[["date"]][1]
ts_start <- c(year(date_start), month(date_start))
inf_head_ts <- ts(inf_head, start = ts_start, frequency = 12)

wshr_obj <- wav_args_wshr(list(
wavelet = c("haar", "d4", "d6", "d8", "s8"),
n.level = 1:4
))

core_wavelet <- wav_smooth(inf_head, wshr_obj)

core_wavelet2 <- wcore_table(core_wavelet, inf_head_ts)

wcore_summary(core_wavelet2, inf_head)
}
\seealso{
\code{\link{wcore_table}},
\code{\link{wav_smooth}}, \code{\link{mavc}},
\code{\link{no_bias2}}.
}
