% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecastOOP.R
\name{wcore_table}
\alias{wcore_table}
\title{Insert Date Column}
\usage{
wcore_table(x, ts)
}
\arguments{
\item{x}{A tibble generated by the function \code{\link{wav_smooth}}.}

\item{ts}{A time series vector.}
}
\value{
A tibble.
}
\description{
Inserts a data column in an object generated by the function \code{\link{wav_smooth}}.
}
\examples{
library(lubridate)

wshr_obj <- wav_args_wshr(list(
  wavelet = c("haar", "d4", "d6", "d8", "s8"),
  n.level = 1:4
))

inf_head <- coreinf_br[["ipca"]]

date_start <- coreinf_br[["date"]][1]
ts_start <- c(year(date_start), month(date_start))
inf_head_ts <- ts(inf_head, start = ts_start, frequency = 12)

core_wavelet <- wav_smooth(inf_head, wshr_obj)

wcore_table(core_wavelet, inf_head_ts)
}
