% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecastOOP.R
\name{model_est_single}
\alias{model_est_single}
\title{Direct Autoregressive Forecast Model}
\usage{
model_est_single(y, h, lags)
}
\arguments{
\item{y}{A vector or a time series.}

\item{h}{An integer, the horizon from which \code{y} is being forecasted.}

\item{lags}{A data frame with lags to begin and to stop. These
lags can be indirectly obtaneid by the function \code{lags} as example
below shows.}
}
\value{
A list of the following components:
\item{model}{A lm object.}
\item{n_obs}{Number of observations.}
\item{h}{The horizon.}
}
\description{
Estimates a direct autoregressive forecast model.
}
\details{
Condiser the following regression:
\itemize{
\item \eqn{y_{i,t+h} = a_0 + y_{i,t + 1 - pstart} + ... + y_{i,t + 1 - pend}  + e_{i,t + h}}.
}
\code{model_est_single} estimates the above regression for all indexes in the
data frame \code{lags} (\code{p_start} \eqn{\ge 1}).

Details section for the function \code{\link{model_est}} provides
more information. Forecasting obtainded by the direct autoregressive model
can be compared with forecasting where lags of the core inflation
measure are considered in the model as in \code{\link{model_est}} function.
}
\examples{
x <- arima.sim(n = 30, list(ar = c(0.8897, -0.4858)))
p <- unique(lags(2, 1)[, 1:2])
model_est_single(x, 2, p)
}
\seealso{
\code{\link{model_est}} \code{\link{model_fcast}}
}
